chore: update to jdk-23

+++ src/main/java/cheshire/io_uring_sqe.java
@@ -146,7 +146,7 @@ public final class io_uring_sqe {
 	};
 
 	public static long getOff(MemorySegment data) {
-		return (long) data.get(ValueLayout.JAVA_LONG, offOffset);
+		return data.get(ValueLayout.JAVA_LONG, offOffset);
 	};
 
 	public static void setOff(MemorySegment data, long value) {
@@ -154,7 +154,7 @@ public final class io_uring_sqe {
 	};
 
 	public static long getAddr(MemorySegment data) {
-		return (long) data.get(ValueLayout.JAVA_LONG, addrOffset);
+		return data.get(ValueLayout.JAVA_LONG, addrOffset);
 	};
 
 	public static void setAddr(MemorySegment data, long value) {
@@ -170,7 +170,7 @@ public final class io_uring_sqe {
 	};
 
 	public static int getRwFlags(MemorySegment data) {
-		return (int) data.get(ValueLayout.JAVA_INT, anon522Offset);
+		return data.get(ValueLayout.JAVA_INT, anon522Offset);
 	};
 
 	public static void setRwFlags(MemorySegment data, int value) {
@@ -178,7 +178,7 @@ public final class io_uring_sqe {
 	};
 
 	public static int getMsgFlags(MemorySegment data) {
-		return (int) data.get(ValueLayout.JAVA_INT, anon522Offset);
+		return data.get(ValueLayout.JAVA_INT, anon522Offset);
 	};
 
 	public static void setMsgFlags(MemorySegment data, int value) {
@@ -186,7 +186,7 @@ public final class io_uring_sqe {
 	};
 
 	public static int getTimeoutFlags(MemorySegment data) {
-		return (int) data.get(ValueLayout.JAVA_INT, anon522Offset);
+		return data.get(ValueLayout.JAVA_INT, anon522Offset);
 	};
 
 	public static void setTimeoutFlags(MemorySegment data, int value) {
@@ -194,7 +194,7 @@ public final class io_uring_sqe {
 	};
 
 	public static int getAcceptFlags(MemorySegment data) {
-		return (int) data.get(ValueLayout.JAVA_INT, anon522Offset);
+		return data.get(ValueLayout.JAVA_INT, anon522Offset);
 	};
 
 	public static void setAcceptFlags(MemorySegment data, int value) {
@@ -202,7 +202,7 @@ public final class io_uring_sqe {
 	};
 
 	public static int getCancelFlags(MemorySegment data) {
-		return (int) data.get(ValueLayout.JAVA_INT, anon522Offset);
+		return data.get(ValueLayout.JAVA_INT, anon522Offset);
 	};
 
 	public static void setCancelFlags(MemorySegment data, int value) {
@@ -218,7 +218,7 @@ public final class io_uring_sqe {
 	};
 
 	public static short getBufIndex(MemorySegment data) {
-		return (short) data.get(ValueLayout.JAVA_SHORT, bufIndexOffset);
+		return data.get(ValueLayout.JAVA_SHORT, bufIndexOffset);
 	};
 
 	public static void setBufIndex(MemorySegment data, short value) {
@@ -234,7 +234,7 @@ public final class io_uring_sqe {
 	};
 
 	public static int getFileIndex(MemorySegment data) {
-		return (int) data.get(ValueLayout.JAVA_INT, fileIndexOffset);
+		return data.get(ValueLayout.JAVA_INT, fileIndexOffset);
 	};
 
 	public static void setFileIndex(MemorySegment data, int value) {
@@ -242,7 +242,7 @@ public final class io_uring_sqe {
 	};
 
 	public static long getAddr3(MemorySegment data) {
-		return (long) data.get(ValueLayout.JAVA_LONG, addr3Offset);
+		return data.get(ValueLayout.JAVA_LONG, addr3Offset);
 	};
 
 	public static void setAddr3(MemorySegment data, long value) {
@@ -250,7 +250,7 @@ public final class io_uring_sqe {
 	};
 
 	public static long getPad2(MemorySegment data) {
-		return (long) data.get(ValueLayout.JAVA_LONG, pad2Offset);
+		return data.get(ValueLayout.JAVA_LONG, pad2Offset);
 	};
 
 	public static void setPad2(MemorySegment data, long value) {
